Move notification permission request to splash
