+++ app/src/main/java/io/gnosis/safe/ui/StartActivity.kt
@@ -3,14 +3,11 @@ package io.gnosis.safe.ui
 import android.content.Context
 import android.content.Intent
 import android.graphics.Rect
-import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import android.text.TextUtils
 import android.view.View
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.view.marginLeft
 import androidx.core.view.marginRight
 import androidx.lifecycle.Lifecycle
@@ -68,8 +65,6 @@ class StartActivity : BaseActivity(), SafeOverviewNavigationHandler, AppStateLis
         ToolbarSafeOverviewBinding.bind(binding.toolbarContainer.root)
     }
 
-    private lateinit var pushNotificationPermissionLauncher: ActivityResultLauncher<String>
-
     private val handler = Handler(Looper.getMainLooper())
 
     var comingFromBackground = true
@@ -77,10 +72,6 @@ class StartActivity : BaseActivity(), SafeOverviewNavigationHandler, AppStateLis
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-
-        pushNotificationPermissionLauncher = registerForActivityResult(
-            ActivityResultContracts.RequestPermission()
-        ) {}
 
         navController = Navigation.findNavController(this@StartActivity, R.id.nav_host)
 
@@ -427,18 +418,11 @@ class StartActivity : BaseActivity(), SafeOverviewNavigationHandler, AppStateLis
     private fun onAppUpdated() {
         showWhatsNew()
         settingsHandler.showWhatsNew = false
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            requestPushPermission()
-        }
     }
 
     private fun showWhatsNew() {
         //TODO: uncomment if whats new screen should be shown
         //navController.navigate(R.id.whatsNewDialog)
-    }
-
-    private fun requestPushPermission() {
-        pushNotificationPermissionLauncher.launch(android.Manifest.permission.POST_NOTIFICATIONS)
     }
 
     private fun navigateToShareSafeDialog() {
